Fix event listener crash
The MI operation cancel/close calls are asynchronuous for async
subscriptions. The issue is that we're not waiting for those
operations to complete when tearing down event watchers.

For this reason, we may get event callbacks after deallocating
those objects, which will lead to invalid memory access and
crashes.

This can be avoided by waiting for the MI operations to be canceled
before going further with the teardown. After MI operations are
canceled, a cancel notification is emitted and the "more_results"
is set to False. We'll just wait for that.