EventWatcher: avoid using weakref conn
If the _Connection object gets cleaned up before the event watcher,
MI crashes. This happens even if the _Connection destructor closes
the event watcher.

It's difficult to say what exactly MI is trying to access from the
crash dump alone.

One workaround is to avoid using weak references for the connection
object. This basically means that we have a circular dependency as the
connection object holds a reference to the event watcher callback.
This dependency gets cleared when calling the watcher .close method,
allowing the two objects to be cleaned up.

In short, with or without this patch, the watcher .close method
must be called when we're done with it (unless we're running a
daemon). Without this patch, if this call is omitted, the process
can crash. With it, it would lead to a memory leak due to the
ever growing queue of received events.